requirements.txt
# Core dependencies
PyMuPDF==1.23.8
openai-whisper==20231117
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Utilities
tqdm==4.66.1
numpy==1.24.3 